requirements.txt
pandas==1.1.5
requests==2.26.0
pyyaml==5.4.1
python-slugify==5.0.2 
pytables==3.6.1
folium== 0.12.1